Move indexing jobs into background (HACKILY).
