+++ internal/codeintel/autoindexing/background/dependencies/dependency_indexing_scheduler.go
@@ -1,4 +1,4 @@
-package indexing
+package dependencies
 
 import (
 	"context"
