+++ internal/codeintel/autoindexing/background/dependencies/main_test.go
@@ -1,4 +1,4 @@
-package indexing
+package dependencies
 
 import (
 	"flag"
