+++ internal/codeintel/autoindexing/background/dependencies/observability.go
@@ -1,4 +1,4 @@
-package indexing
+package dependencies
 
 import (
 	"fmt"
